fix most tests

+++ contracts/interfaces/IChainlink.sol
@@ -3,7 +3,7 @@ pragma solidity >=0.7.2;
 pragma experimental ABIEncoderV2;
 
 interface IChainlink {
-    function decimals() external view returns (uint256);
+  function decimals() external view returns (uint8);
 
-    function latestAnswer() external view returns (uint256);
+  function latestAnswer() external view returns (int256);
 }
